requirements.txt
annotated-types==0.7.0
anyio==4.13.0
asgiref==3.11.1
certifi==2026.4.22
charset-normalizer==3.4.7
cloudinary==1.44.2
distro==1.9.0
dj-database-url==3.1.2
Django==4.2.30
django-cloudinary-storage==0.3.0
exceptiongroup==1.3.1
groq==1.2.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.13
pillow==12.2.0
psycopg2-binary==2.9.12
pydantic==2.13.3
pydantic_core==2.46.3
python-dotenv==1.2.2
requests==2.33.1
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.5
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
whitenoise==6.8.2
